Moved BaseToObjectValueWriter out of public API
