+++ community/cypher/cypher/src/main/java/org/neo4j/cypher/internal/javacompat/ValueToObjectSerializer.java
@@ -24,7 +24,7 @@ import org.neo4j.cypher.internal.compatibility.v3_3.runtime.GeographicPoint;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Relationship;
 import org.neo4j.graphdb.spatial.Point;
-import org.neo4j.helpers.BaseToObjectValueWriter;
+import org.neo4j.kernel.impl.util.BaseToObjectValueWriter;
 import org.neo4j.kernel.impl.core.NodeManager;
 
 public class ValueToObjectSerializer extends BaseToObjectValueWriter<RuntimeException>
